[util]: Multiple parameters & multiple lines
param.sh now properly outputs parameters. The previous implementation would
be considered a bug if input from stdin reading from say a device as it
required an EOF before start pocessing the parameters. Additionaly only the
first one with the required name would be found.

This version might be slower but if that becomes annoying, script should be
rewritten in a better script-language, but function must remain.

Category: bugfix